requirements.txt
discord.py>=2.6.2
python-dotenv>=1.1.1
asyncpg>=0.29.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
